Rework to use differential evolution.
This is a popular approach for filter design, and it certainly works
much better than the simple GA (it produces results in about 20 minutes
that are better than the simple GA running for an afternoon). Better
parallelisation also helps!

On the other hand, the results it finds don't seem to cross over very
effectively, which suggests the filter representation isn't ideal...