Remove the zero-crossing hysteresis.
It's now detecting crossings of the boundary between 0 and -1, rather
than splitting the difference between 1 and -1.

This only makes a difference to the output very occasionally (the first
example in fawlty.s16 shows up after 140,000 crossings), but it makes a
tiny improvement to the error rates in all my samples, even managing to
squeeze a few thousand samples' more audio out of some of the worst
examples.