Read the actual composite value from a third log file.
The previous approach wasn't accurately computing the composite value,
because it only knows the magnitudes and not the phases of the bins --
if the chroma and luma don't have the same phase, then adding their
magnitudes doesn't give the composite magnitude. We could log the phases
as well, but it's cheaper to just do a third run and log the composite
value.